Fix Xmonad config to live in ~/.xmonad/xmonad.hs
I was having trouble with recompilation when it was part of the nix
configuration. I just couldn't recompile without "nixos-rebuild
switch" which feels too heavyweight. Not only that, but mod-Q doesn't
work in that scenario.  Now it should work, and I have a proper shared
library instead of jankily concatenating the shared config with main!